Коммит 2 по проектной работе №1 (с исправлениями)
